refactor: isolate stair geometry
